// src/api/appApiClient.ts
export class SeachModel implements ISeachModel {
    skip?: number;
    size?: number;
    keyword?: string | undefined;

    constructor(data?: ISeachModel) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.skip = _data["skip"];
            this.size = _data["size"];
            this.keyword = _data["keyword"];
        }
    }

    static fromJS(data: any): SeachModel {
        data = typeof data === 'object' ? data : {};
        let result = new SeachModel();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["skip"] = this.skip;
        data["size"] = this.size;
        data["keyword"] = this.keyword;
        return data;
    }
}

export interface ISeachModel {
    skip?: number;
    size?: number;
    keyword?: string | undefined;
}

export class Test implements ITest {
    id?: number;
    createdOn?: Date | undefined;
    lastModifyOn?: Date | undefined;
    isDeleted?: boolean;
    name?: string | undefined;
    age?: number;
    address?: string | undefined;
    country?: string | undefined;

    constructor(data?: ITest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.createdOn = _data["createdOn"] ? new Date(_data["createdOn"].toString()) : <any>undefined;
            this.lastModifyOn = _data["lastModifyOn"] ? new Date(_data["lastModifyOn"].toString()) : <any>undefined;
            this.isDeleted = _data["isDeleted"];
            this.name = _data["name"];
            this.age = _data["age"];
            this.address = _data["address"];
            this.country = _data["country"];
        }
    }

    static fromJS(data: any): Test {
        data = typeof data === 'object' ? data : {};
        let result = new Test();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["createdOn"] = this.createdOn ? this.createdOn.toISOString() : <any>undefined;
        data["lastModifyOn"] = this.lastModifyOn ? this.lastModifyOn.toISOString() : <any>undefined;
        data["isDeleted"] = this.isDeleted;
        data["name"] = this.name;
        data["age"] = this.age;
        data["address"] = this.address;
        data["country"] = this.country;
        return data;
    }
}

export interface ITest {
    id?: number;
    createdOn?: Date | undefined;
    lastModifyOn?: Date | undefined;
    isDeleted?: boolean;
    name?: string | undefined;
    age?: number;
    address?: string | undefined;
    country?: string | undefined;
}